requirements.txt
streamlit
pandas
requests
Pillow
supabase